AerugoHal: Renamed `SystemHal` to `AerugoHal` and simplifed it's crate
