+++ src/lib.rs
@@ -33,9 +33,7 @@ pub use self::boolean_condition::{
 pub use self::event::{EventEnabler, EventId};
 pub use self::message_queue::{MessageQueueHandle, MessageQueueStorage};
 pub use self::tasklet::{TaskletConfig, TaskletStorage};
-pub use aerugo_hal::system_hal::SystemHardwareConfig;
-
-pub use aerugo_hal::{time, Duration, Instant};
+pub use aerugo_hal::{time, Duration, Instant, SystemHardwareConfig};
 
 #[cfg(feature = "use-aerugo-cortex-m")]
 pub(crate) use aerugo_cortex_m as arch;
